% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundPlots.R
\name{abundPlots}
\alias{abundPlots}
\title{Make taxonomic abundance plot}
\usage{
abundPlots(
  d,
  family = family,
  genus = genus,
  species = NULL,
  genSpp = NULL,
  colourSet = "Set1",
  scaleYtext = c(1, 1, 1),
  keepSpp = TRUE
)
}
\arguments{
\item{d}{Data frame or tibble}

\item{family}{Family column}

\item{genus}{Genus column}

\item{species}{Species column - can also use genSpp column (see below)}

\item{genSpp}{GenSpp column (optional - create using \code{\link{makeGenSpp}})}

\item{colourSet}{Colour set from ColourBrewer (default = "Set1")}

\item{scaleYtext}{Scaling factor for y-axis text (default = c(1,1,1))}

\item{keepSpp}{Keep morphospecies? (default = TRUE)}
}
\value{
A \code{ggarrange} (\code{ggplot}) object, showing bee abundances split up by species, genus, and family
}
\description{
Function to make bee taxonomic abundance plots (species, genus, family level counts) from data frame.
}
\examples{
dat <- data.frame(f=c('Apidae','Apidae','Apidae','Colletidae','Andrenidae','Andrenidae'),
    g=c('Bombus','Bombus','Apis','Hylaeus','Lasioglossum','Lasioglossum'),
    s=c('rufocinctus','rufocinctus','mellifera','annulatus','zonulum','spp.'))
abundPlots(dat,f,g,s,keepSpp=FALSE)
abundPlots(dat,f,g,s,keepSpp=TRUE)
}
