% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundTable.R
\name{abundTable}
\alias{abundTable}
\title{Create a \code{kable} abundance table}
\usage{
abundTable(d, col1, col2, totals = FALSE)
}
\arguments{
\item{d}{Data frame or tibble}

\item{col1}{First column to use (e.g. species)}

\item{col2}{Second column to use (e.g. site)}

\item{totals}{Should column and row totals be added?}
}
\value{
A \code{kable} table
}
\description{
#Function to make abundance kable table across two marginal variables (col1, col2)
}
\examples{
NULL
}
