% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmy2date.R
\name{dmy2date}
\alias{dmy2date}
\title{Create Date column}
\usage{
dmy2date(d, day, month, year, newName = "date", dateFormat = "\%d_\%m_\%Y")
}
\arguments{
\item{d}{Data frame}

\item{day}{Day column}

\item{month}{Month column}

\item{year}{Year column}

\item{newName}{Name for new date column (default = "date")}

\item{dateFormat}{Format for date object (default = "%d_%m_%Y", see \code{strptime})}
}
\value{
A data frame
}
\description{
Converts day,month,year to a Date column (\code{as.Date})
}
\examples{
dat <- data.frame(a=letters[1:2],d=c(10,12),m=c(6,12),y=c(1988,2012))
dmy2date(dat,d,m,y,'newDate')
}
