% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srtools.R
\name{en.ern}
\alias{en.ern}
\title{en_ern}
\usage{
en.ern(xmlrecord)
}
\arguments{
\item{xmlrecord}{a <record>...</record> entry from xml}
}
\value{
tibble with <ern> column
}
\description{
get `electronic-resource-num` from endnote records
}
