% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genFamTable.R
\docType{data}
\name{genFamTable}
\alias{genFamTable}
\title{Familes of common bee genera}
\format{
A data frame with 49 rows and 2 variables:
\describe{
  \item{family}{family}
  \item{genus}{genus}
}
}
\usage{
genFamTable
}
\description{
A dataset containing two columns, matching common bee genera to families
}
\keyword{datasets}
