% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLabels.R
\name{makeLabels}
\alias{makeLabels}
\title{Print insect labels}
\usage{
makeLabels(inputPath, csvPaths, outputPath, sepChar = rep(",", 3))
}
\arguments{
\item{inputPath}{Input file with number of labels needed (csv)}

\item{csvPaths}{Paths to site and trap tables (vector with 2 paths - csv)}

\item{outputPath}{Output path (docx)}

\item{sepChar}{Separator characters for \code{inputPath}, \code{site}, and
\code{trap} table csvs (default: rep(',',3))}
}
\description{
Function to print insect labels from a set of csv files. Written
  for labels used in the U of C insect collection and Galpern lab.
}
\details{
\strong{\code{inputPath}} is a path to a csv with the following
  column headers:

  \emph{BTID} (trapping ID), \emph{N} (number of labels needed)

  \strong{\code{csvPaths[1]}} is a path to the \code{site} csv with the
  following column headers:

  \emph{BLID}, \emph{lat}, \emph{lon}, \emph{locality}, \emph{elevation},
  \emph{country}, \emph{province}

  \strong{\code{csvPaths[2]}} is a path to the \code{trap} csv with the
  following column headers:

  \emph{BTID}, \emph{BLID}, \emph{collector}, \emph{trapType},
  \emph{startYear}, \emph{startMonth} (numeric), \emph{startDay},
  \emph{endYear}, \emph{endMonth} (numeric), \emph{endDay}, \emph{lonTrap},
  \emph{latTrap}, \emph{elevTrap}

  If \emph{lonTrap}, \emph{latTrap}, or \emph{elevTrap} values are missing
  from \code{csvPaths[2]}, values will be filled using \code{site} csv.

  \strong{\code{output[1]}} is a path to a \emph{docx} file that will be
  printed. If more than 1 page of labels is required, additional files will
  be created with numbers appended before the file extension (e.g.
  \emph{labels_01.docx}, \emph{labels_02.docx})
}
\examples{
# Make csv files for labels

# 2 trapping sessions, with 20 labels each
input <- data.frame(BTID = c('10069-1-A-2015','51273-0-B-2019'),
                    N = c(20,20))

# Site table
site <- data.frame(BLID=c(10069,51273),lat=c(50,51),lon=c(-114,-113),
                   locality=c('Near Town','Near Mountains'),
                   elevation=c(800,1200),country=rep('CANADA',2),
                   province=rep('Alberta',2))

#Trap table - netting has only a single day on label, while blue vane has a
date range
trap <- data.frame(BTID=c('10069-1-A-2015','51273-0-B-2019'),
                   BLID=c(10069,51273),collector=c('D. Smith','L. Robbins'),
                   trapType=c('Blue Vane','Netting'),
                   startYear=c(2012,2021),startMonth=c(6,7),startDay=c(10,20),
                   endYear=c(2012,NA),endMonth=c(6,NA),endDay=c(15,NA),
                   lonTrap=c(-114.5,-113.1),latTrap=c(51,52.3),
                   elevTrap=c(815,1231))

#Write csvs to file
write.csv(input,'inputcsv.csv')
write.csv(site,'sitecsv.csv')
write.csv(trap,'trapcsv.csv')

#Create labels
makeLabels(inputPath = 'inputcsv.csv',
           csvPaths = c('sitecsv.csv','trapcsv.csv'),
           outputPath = 'labelSheet.docx'
)
}
