% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propAbundPlots.R
\name{propAbundPlots}
\alias{propAbundPlots}
\title{propAbundPlots}
\usage{
propAbundPlots(
  d,
  s,
  N,
  textYRange = c(0.5, 0.9),
  by = NULL,
  Ncol = NA,
  Nrow = NA
)
}
\arguments{
\item{d}{Data frame or tibble}

\item{s}{Species label column}

\item{N}{Species count column}

\item{textYRange}{Numeric vector with proportional y-range for plot text (default: c(0.5, 0.9))}

\item{by}{Faceting column - usually site (default: NULL)}

\item{Ncol}{Number of columns, if using facets}

\item{Nrow}{Number of rows, if using facets}
}
\value{
A \code{ggplot} object
}
\description{
#Make proportional abundance bar-plots, with a rank-order smoothing line
}
\examples{
data(dune,package='vegan')

library(tidyverse)

dune <- dune \%>\% mutate(site=1:nrow(.)) \%>\%
  pivot_longer(cols = -site,names_to='species',values_to='counts')

  propAbundPlots(dune,s = species,N = counts)
  filter(dune,site<10) \%>\% #First 9 sites
    propAbundPlots(s = species,N = counts, by = site, Nrow=3, Ncol=3)

}
