% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteRarePlots.R
\name{siteRarePlots}
\alias{siteRarePlots}
\title{Site-based rarefaction plots}
\usage{
siteRarePlots(
  d,
  Ncol = NA,
  Nrow = NA,
  measType = "Chao1",
  textRange = c(0.1, 0.4),
  seMult = 1,
  rowOrder = "asis"
)
}
\arguments{
\item{d}{Matrix of species abundance (rows = sites, cols = species)}

\item{Ncol}{Number of columns for facets}

\item{Nrow}{Number of rows for facets}

\item{measType}{Diversity predictor ('Chao1' or 'none')}

\item{textRange}{Proportion upper/lower bounds for text display. Can be 2 values overall, or 2 x N sites (min1,max1,min2,max2,...,minN,maxN).}

\item{seMult}{Multiplier for SE (default = 1)}

\item{rowOrder}{Should sites in facets be ordered by row order ('asis'), observed diversity ('Ndiv'), or # of samples ('Nsamp')?}
}
\value{
A \code{ggplot} object
}
\description{
#Function to make "Linc-style" rarefaction plots using \pgk{vegan}
}
\examples{
data(dune,package='vegan')

#Site-by-site
siteRarePlots(dune,Ncol = 5,Nrow=4,measType = 'Chao1',seMult = 2)

#Overall
oDune <-t(as.matrix(colSums(dune)))
siteRarePlots(oDune,Ncol = 1,Nrow=1,measType = 'Chao1',seMult = 2)
}
