% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wideTable.R
\name{wideTable}
\alias{wideTable}
\title{Create a wide kable table}
\usage{
wideTable(
  d,
  countBy = {
     {
         countBy
     }
 },
  countNum = {
     {
         countNum
     }
 },
  nsplit = NA,
  colNames = c("countBy", "count"),
  caption = "",
  returnDF = FALSE,
  ...
)
}
\arguments{
\item{d}{Data frame or tibble}

\item{nsplit}{How many splits should be made?}

\item{colNames}{Change column names}

\item{caption}{Table caption}

\item{returnDF}{Return results as a data frame?}

\item{...}{Other arguments for \code{kable_styling}}
}
\value{
A kable table
}
\description{
Function to make a "wide kable table", with column names replicated across the page
}
\examples{
NULL
}
