% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteRarePlots.R
\name{siteRarePlots}
\alias{siteRarePlots}
\title{Site-based rarefaction plots}
\usage{
siteRarePlots(
  d,
  Ncol = NA,
  Nrow = NA,
  measType = "both",
  textRange = c(0.1, 0.4),
  seMult = 1,
  rowOrder = "asis"
)
}
\arguments{
\item{d}{Matrix of species abundance (rows = species, cols = sites)}

\item{Ncol}{Number of columns for facets}

\item{Nrow}{Number of rows for facets}

\item{measType}{Type of diversity predictor ('both','Chao1','ACE','none')}

\item{textRange}{Proportion upper/lower bounds for text display. Can be 2 values overall, or 2 x N sites (min1,max1,min2,max2,...,minN,maxN).}

\item{seMult}{Multiplier for SE (default = 1)}

\item{rowOrder}{Should sites in facets be ordered in row order ('asis'), by diversity ('estDiv'), or # of samples ('Nsamp')?}
}
\value{
A \code{ggplot} object
}
\description{
#Function to make "Linc-style" rarefaction plots using \pgk{vegan}
}
\examples{
NULL
}
